% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cairo5-matched-sequencing}
\alias{cairo5-matched-sequencing}
\title{This}
\value{
Number of distinct and mutant reads in white blood cells
and cell-free DNA tissues for each patient's mutations (each row)
cairo5-matched-sequencing
}
\description{
Processed matched sequencing output from the CAIRO5 clinical trial.
The data in this sheet is presented in this paper: 
The cohort and analyses are described here: 
\url{https://pubmed.ncbi.nlm.nih.gov/36534496/}
}
\details{
The data in this csv file (inst/extdata/cairo5-matched-sequencing.csv)
is processed according to inst/script/crcseq/R to create crcseq.rda 
that is fed through plasmut to estimate the probability
of tumor specific alterations
}
