% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumorprobability.R
\name{compute_p_tumor}
\alias{compute_p_tumor}
\title{Compute the likelihood of a given plasma mutation being somatic in origin}
\usage{
compute_p_tumor(
  data,
  rho_ctc_parameter = 0,
  a_w_somatic = 1,
  b_w_somatic = 10,
  a_p_somatic = 1,
  b_p_somatic = 10,
  a_theta = 1,
  b_theta = 10,
  prior.odds = 1,
  chains = 2,
  adapt = 2000,
  iter = 1e+05,
  nthin = 10
)
}
\arguments{
\item{data}{is a list with 4 named items (np = total distinct reads plasma, yp = mutant reads plasma, nw = total distinct reads wbc, yw = mutant reads wbc)}

\item{rho_ctc_parameter}{fraction of cells in buffy coat layer expected to be circulating tumor cells, default 0. If CTC suspected, 0.001 is a good starting point.}

\item{a_w_somatic}{shape beta parameter for informing wbc maf in somatic CTC model}

\item{b_w_somatic}{rate beta parameter for informing wbc maf in somatic CTC model}

\item{a_p_somatic}{shape beta parameter for informating plasma maf in somatic CTC and No CTC model}

\item{b_p_somatic}{rate beta parameter for informating plasma maf in somatic CTC and No CTC model}

\item{a_theta}{shape beta parameter for informing plasma and wbc maf in hematopoietic model}

\item{b_theta}{rate beta parameter for informing plasma and wbc maf in hematopoietic model}

\item{prior.odds}{P(mutation being somatic) / P(mutation being hematopoietic)}

\item{chains}{number of monte carlo chains, default 2}

\item{adapt}{number of iterations for burn in of each monte carlo chain, default 2000}

\item{iter}{number of iterations for each chain, default 1e5}

\item{nthin}{thinning value, take every nth value of chain, default 10}

\item{samples}{# of samples for Monte Carlo integration}
}
\value{
list with two named vectors - bf (bayes factor) and p (probbability of tumor-specific) for all mutations supplied into function
}
\description{
Compute the likelihood of a given plasma mutation being somatic in origin
}
\examples{
compute_p_tumor(list(np=100, yp=5, nw=100, yw=10))
}
