% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crcseq}
\alias{crcseq}
\title{This data is an example dataset to show how to use the package}
\value{
An example DNA sequencing dataset of matched plasma 
and wbc colorectal cancer samples
crcseq
}
\description{
A cohort of metastatic colorectal cancer patients whose plasma and buffy 
coat were sequenced as part of the CAIRO5 trial. 
The cohort and analyses are described here: 
\url{https://pubmed.ncbi.nlm.nih.gov/36534496/}
}
