% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfvar-plot.R
\name{geom_cfvar}
\alias{geom_cfvar}
\title{visualize probability of mutation being somatic against maf of wbc and plasma}
\usage{
geom_cfvar(df)
}
\arguments{
\item{df}{dataframe with somatic probabilities computed by mutchips, wbc var type categorized by mutchips, and wbcmaf and cfdna maf computed as specificed in vignette}
}
\value{
a ggplot2 plot that can be further customized
}
\description{
visualize probability of mutation being somatic against maf of wbc and plasma
}
