% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorodds.R
\name{lik.somatic}
\alias{lik.somatic}
\title{Compute the likelihood of a given plasma mutation being somatic in origin}
\usage{
lik.somatic(y.p, n.p, y.w, n.w, thetas, p.theta)
}
\arguments{
\item{y.p}{# of distinct mutant reads in plasma}

\item{n.p}{# of distinct reads in plasma}

\item{y.w}{# of distinct mutant reads in wbc}

\item{n.w}{# of distinct reads in wbc}

\item{thetas}{samples from a beta distribution}

\item{p.theta}{density of beta computed at thetas (i.e. dbeta(thetas, a, b))}
}
\value{
the likelihood of the mutation being somatic in origin
}
\description{
Compute the likelihood of a given plasma mutation being somatic in origin
}
