% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{model_w}
\alias{model_w}
\title{Estimate the marginal likelihood that mutations in buffy coat and cfDNA reflect CH or correspond to germline mutations.}
\usage{
model_w(dat, params)
}
\description{
If germline, the allele frequency should be 50%.  The prior should be diffuse enough to handle CHIP mutations (potentially << 50%).
}
