% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{plasma_somatic}
\alias{plasma_somatic}
\title{Estimate the marginal likelihood that variants identified in cell-free DNA are derived from tumor cells (ctDNA-derived)}
\usage{
plasma_somatic(dat, params)
}
\description{
Estimate the marginal likelihood that variants identified in cell-free DNA are derived from tumor cells (ctDNA-derived)
}
