% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{plasmut}
\alias{plasmut}
\title{Bayesian models for estimating the origin of a sequenced DNA fragment}
\description{
The plasmut package provides a Bayesian importance sampling based approach 
to estimate the liklihood of a mutation arising from clonal hematopoiesis
or tumor
}
