% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorodds.R
\name{posterior.odds}
\alias{posterior.odds}
\title{Posterior Odds of a mutation being somatic}
\usage{
posterior.odds(bayes.factors, prior.odds = 1)
}
\arguments{
\item{bayes.factors}{is a vector of bayes factors computed from the function bayes.factors() provided in this package}
}
\value{
a vector of posterior odds
}
\description{
Posterior Odds of a mutation being somatic
}
\examples{
posterior.odds(2, 1)
}
