% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorodds.R
\name{posterior.probability}
\alias{posterior.probability}
\title{Posterior Probability of mutation being somatic (tumor specific) computed from Posterior Odds}
\usage{
posterior.probability(po)
}
\arguments{
\item{po}{a vector of posterior odds}
}
\value{
a vector of posterior probabilities
}
\description{
Posterior Probability of mutation being somatic (tumor specific) computed from Posterior Odds
}
\examples{
posterior.probability(1)
}
