% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbcvariant-handling.R
\name{stratify_wbc_vars}
\alias{stratify_wbc_vars}
\title{handle mutations with a non-zero white blood cell mutant read count}
\usage{
stratify_wbc_vars(
  pdmr,
  pdr,
  wdmr,
  wdr,
  thresh = NULL,
  data = NULL,
  cols = NULL
)
}
\arguments{
\item{pdmr}{vector of distinct mutant reads in plasma for different mutations}

\item{pdr}{vector of distinct reads in plasma for different mutations}

\item{wdmr}{vector of distinct mutant reads in wbc for different mutations}

\item{wdr}{vector of distinct reads in wbc for different mutations}

\item{data}{data frame with the pdmr, pdr, wdmr, and wdr information as column and rows correspond to different mutations}

\item{cols}{vector of column names to use (cols[1] refers to pdmr, etc.)}
}
\value{
the vector representing the clusters to which the white blood cell variants belong to
}
\description{
handle mutations with a non-zero white blood cell mutant read count
}
