% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbcvariant-handling.R
\name{strip.column}
\alias{strip.column}
\title{Strip a column of a dataframe into a vector}
\usage{
strip.column(data, col.idx = 1)
}
\arguments{
\item{data}{a data frame.}

\item{col.idx}{the column name or column number}
}
\value{
a vector of the selected column
}
\description{
Strip a column of a dataframe into a vector
}
