% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{importance_sampler}
\alias{importance_sampler}
\title{Importance sampler to estimate marginal likelihoods and Bayes factors}
\usage{
importance_sampler(dat, params, save_montecarlo = TRUE)
}
\description{
Importance sampler to estimate marginal likelihoods and Bayes factors
}
\examples{
param.list <- list(ctc=list(a=1, b=9999),
                   ctdna=list(a=1, b=9),
                   chip=list(a=1, b=9),
                   montecarlo.samples=50e3,
                   prior.weight=0.1)
dat <- tibble(y=c(4, 1),
              n=c(1000, 1000),
              analyte=c("plasma", "buffy coat"),
              mutation="mutA",
              sample_id="id1")
importance_sampler(dat, param.list)
}
