% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{importance_samples}
\alias{importance_samples}
\title{Generate importance samples}
\usage{
importance_samples(y, n, a, b, S, prior.weight, model, dna_source)
}
\description{
Generate importance samples
}
