% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorodds.R
\name{posterior.odds}
\alias{posterior.odds}
\title{Compute the likelihood of a given plasma mutation being somatic in origin}
\usage{
posterior.odds(
  pdmr,
  pdr,
  wdmr,
  wdr,
  data = NULL,
  cols = NULL,
  samples = 1e+06,
  prior.odds = 1
)
}
\arguments{
\item{pdmr}{vector of distinct mutant reads in plasma for different mutations}

\item{pdr}{vector of distinct reads in plasma for different mutations}

\item{wdmr}{vector of distinct mutant reads in wbc for different mutations}

\item{wdr}{vector of distinct reads in wbc for different mutations}

\item{data}{data frame with the pdmr, pdr, wdmr, and wdr information as column and rows correspond to different mutations}

\item{cols}{vector of column names to use (cols[1] refers to pdmr, etc.)}

\item{samples}{# of samples for Monte Carlo integration}

\item{prior.odds}{P(mutation being somatic) / P(mutation being hematopoietic)}
}
\value{
the posterior odds of the mutation being somatic
}
\description{
Compute the likelihood of a given plasma mutation being somatic in origin
}
