% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{wbc_somatic}
\alias{wbc_somatic}
\title{Estimate the marginal likelihood of observing somatic mutations from CTCs present in buffy coat}
\usage{
wbc_somatic(dat, params)
}
\arguments{
\item{dat}{tibble containing elements `y`and `n`.  The `y` and `n` vectors should be named}

\item{params}{a list with named elements that must include the following:
`gamma`: weight given to heavier-tailed prior distribution
`a`: prior expectation for number of somatic variants observed in the WBC sequencing data (either by error or from a CTC)
`b`: prior expectation for number of WBCs not containing the variant}
}
\description{
p(y_w | theta_w, n_w, model_S) x p(theta_w| Model_S)
theta_w | model_S  ~ beta(1, 999)  ## sequencing error or CTC
}
\examples{
## asssume we observed y_w=500 and n_w = 1000
## this is clearly germline, but to go through the machinery we would obtain the probability of observing 500 someatic
## variants (y_w) under a model where the prior for theta_w has most of the mass near zero
##    - the importance sampler does not really make sense here as it samples from a distribution that is fatter tailed than the
##      posterior -> it

}
